Combined react ejected with manual react setup- compiling webpack successfully; Pending : dotenv config
