+++ config/env.js
@@ -1,5 +1,5 @@
 'use strict';
-
+require("dotenv").config();
 const fs = require('fs');
 const path = require('path');
 const paths = require('./paths');
