--- src/Components/Authentication/ForgotPassword/index.jsx
@@ -1,57 +1,0 @@
-import React, { useState } from "react"
-import { Form, Button, Card, Alert } from "react-bootstrap"
-import { Link } from "react-router-dom"
-// import { useAuth } from "../../../Utils/Auth"
-
-export default function ForgotPassword() {
-  // const loggedInUser = useAuth()
-  const [error, setError] = useState("")
-  const [message, setMessage] = useState("")
-  const [loading, setLoading] = useState(false)
-
-  async function handleSubmit(e) {
-    e.preventDefault()
-    try {
-      setMessage("")
-      setError("")
-      setLoading(true)
-      // await resetPassword(emailRef.current.value)
-      setMessage("Check your inbox for further instructions")
-    } catch {
-      setError("Failed to reset password")
-    }
-
-    setLoading(false)
-  }
-
-  return (
-    <>
-      <Card>
-        <Card.Body>
-          <h2 className="text-center mb-4">Password Reset</h2>
-          {error && <Alert variant="danger">{error}</Alert>}
-          {message && <Alert variant="success">{message}</Alert>}
-          <Form
-            onSubmit={handleSubmit}
-          >
-            <Form.Group id="email">
-              <Form.Label>Email</Form.Label>
-              <Form.Control type="email"
-                // ref={emailRef}
-                required />
-            </Form.Group>
-            <Button disabled={loading} className="w-100" type="submit">
-              Reset Password
-            </Button>
-          </Form>
-          <div className="w-100 text-center mt-3">
-            <Link to="/authenticate/login">Login</Link>
-          </div>
-        </Card.Body>
-      </Card>
-      <div className="w-100 text-center mt-2">
-        Need an account? <Link to="/authenticate/sign-up">Sign Up</Link>
-      </div>
-    </>
-  )
-}
