--- src/Components/Home/About/index.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const About = () => {
-  return (
-    <h1>   
-      About
-    </h1>
-  )
-}
-
-export default About
