+++ src/Components/Home/About/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const About = () => {
+  return (
+    <h1>   
+      About
+    </h1>
+  )
+}
+
+export default About
