+++ src/Utils/Firebase/index.ts
@@ -1,4 +1,5 @@
 import { initializeApp } from "firebase/app";
+
 import {
   getAuth,
   signInWithRedirect,
@@ -24,15 +25,15 @@ import {
   ObjectToAdd,
   AdditionalInformation,
 } from "../../Types/Users";
-
+// require("dotenv").config();
 const firebaseConfig = {
-  apiKey: process.env["REACT_APP_FIREBASE_API_KEY"],
-  authDomain: process.env["REACT_APP_FIREBASE_AUTH_DOMAIN"],
-  projectId: process.env["REACT_APP_FIREBASE_PROJECT_ID"],
-  storageBucket: process.env["REACT_APP_FIREBASE_STORAGE_BUCKET"],
-  messagingSenderId: process.env["REACT_APP_FIREBASE_MESSAGING_SENDER_ID"],
-  appId: process.env["REACT_APP_FIREBASE_APP_ID"],
-  measurementId: process.env["REACT_APP_FIREBASE_MEASUREMENT_ID"],
+  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
+  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_FIREBASE_APP_ID,
+  measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
 };
 
 export const firebaseApp = initializeApp(firebaseConfig);
