+++ tsconfig.json
@@ -1,29 +1,22 @@
 {
   "compilerOptions": {
-    "target": "es5",
-    "downlevelIteration": true,
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "noFallthroughCasesInSwitch": true,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
-    "jsx": "react-jsx",
-    "noPropertyAccessFromIndexSignature": true
+      "target": "ES5" ,
+      "module": "ESNext" ,
+      "moduleResolution": "node" ,
+      "lib": [
+          "DOM",
+          "ESNext"
+      ] ,
+      "jsx": "react-jsx" ,
+      "noEmit": true ,
+      "isolatedModules": true ,
+      "esModuleInterop": true ,
+      "strict": true ,
+      "skipLibCheck": true ,
+      "forceConsistentCasingInFileNames": true ,
+      "resolveJsonModule": true
   },
   "include": [
-    "src", "src/*.d.ts"
-  ],
-  "exclude": ["node_modules"]
+      "src/**/*"
+  ]
 }
